+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.Mankomania"
         tools:targetApi="31">
             <activity
-                android:name="com.grimschitz.mankomania.BoardScreenActivity"
+                android:name="com.grimschitz.mankomania.BoardScreen.BoardScreenActivity"
                 android:label="@string/title_activity_board_screen"
                 android:theme="@style/Theme.Mankomania.NoActionBar"
                 android:exported="true">
