+++ app/src/main/res/layout/fragment_second.xml
@@ -5,7 +5,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/fragment_Second_Layout"
-        tools:context=".BoardScreenActivity">
+        tools:context=".BoardScreen.BoardScreenActivity">
